chore: split tests into files
